refactor(dispatcher): move package to internal
Implementation detail.
